chore: restructured component Main to App
